requirements.txt
crewai>=0.28.0
langchain-openai>=0.1.0
langchain>=0.1.0
langchain-community>=0.0.20
streamlit>=1.35.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0
